% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{add_iterations_drgemII}
\alias{add_iterations_drgemII}
\title{Add \emph{r}'' iterations to DR-GEM Phase II}
\usage{
add_iterations_drgemII(
  df,
  r_pp,
  iobj,
  query = c(1, 2, 3, 4, 5),
  assign = c(1, 2, 3, 4, 5),
  sigs = NULL,
  markers_field = "markers.scores",
  reprocess = F,
  n_iter = 800,
  n_sub = 600,
  n_dims = 30,
  n_features = 155,
  field = "predicted_clusters",
  align = T,
  params = list(res = 1, min.dist = 0.1, k.param = 10, return.model = F),
  var_features = row.names(iobj)
)
}
\arguments{
\item{df}{assignments across iterations}

\item{r_pp}{number of additional iterations to run}

\item{iobj}{Seurat object to be processed.}

\item{query}{Vector of class labels to subsample.}

\item{reprocess}{Logical, whether to reprocess the Seurat object during each iteration.}

\item{n_iter}{Number of iterations for subsampling.}

\item{n_sub}{Number of subsamples per class.}

\item{n_dims}{Number of principal components to use for PCA.}

\item{n_features}{Number of features to use for variable feature selection.}

\item{field}{Metadata field to use for alignment.}

\item{align}{Logical, whether to align clusters during each iteration.}

\item{params}{List of additional parameters for clustering and UMAP.}
}
\value{
A list of two objects: (1) A dataframe with assigned clusters and
confidence scores for each cell. and (2) A list with the proportion of
change of each cell cluster after the additional \emph{r}'' iterations.
}
\description{
Functionality for adding \emph{r}'' iterations
}
