% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{average.mat.rows}
\alias{average.mat.rows}
\title{Get the average of each matrix}
\usage{
average.mat.rows(m, ids, f = colMeans)
}
\arguments{
\item{m}{matrix}

\item{ids}{identifiers}

\item{f}{averaging function (colMeans)}
}
\description{
Function that gets the average of each column
of a matrix
}
