% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_mse}
\alias{calc_mse}
\title{Calculate Mean Squared Error}
\usage{
calc_mse(X, X_hat)
}
\arguments{
\item{X}{input matrix}

\item{X_hat}{reconstructed matrix}
}
\value{
vector of mean squared error loss
}
\description{
Calculates mean squared error for the the
}
