% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_recon_loss}
\alias{calc_recon_loss}
\title{Calculate Reconstruction Loss}
\usage{
calc_recon_loss(X_scaled, U, n_dims)
}
\arguments{
\item{X_scaled}{Scaled input matrix.}

\item{U}{Matrix of principal components or loadings.}

\item{n_dims}{Number of dimensions to use for reconstruction.}
}
\value{
Vector of reconstruction loss values (mean squared error).
}
\description{
Computes the reconstruction loss of a dimensionality reduction technique,
such as PCA, using mean squared
error between the original and reconstructed data.
}
