% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cap_object}
\alias{cap_object}
\title{Capping the values of an object}
\usage{
cap_object(X, q)

cap_object(X, q)
}
\arguments{
\item{X}{a list-type object: matrix, data.frame, or simple list}

\item{q}{quantile to cap at (e.g. 0.05 is capping at the bottom 5th and top 95th quantiles)}
}
\value{
The input object with values capped to the specified quantiles.
}
\description{
This function caps the values of a numeric object to specific quantiles.

Cap the values of data (matrix, data.frame, or list)
}
