% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cluster_and_viz}
\alias{cluster_and_viz}
\title{Cluster and Visualize Data}
\usage{
cluster_and_viz(
  obj,
  n_dims = 10,
  reduction = "pca",
  clus = T,
  viz = T,
  align = T,
  field1 = "labels",
  field2 = "seurat_clusters",
  field_aligned = "weak_clusters",
  params = list(res = 0.12, min.dist = 0.1, k.param = 20, return.model = F)
)
}
\arguments{
\item{obj}{Seurat object to be processed.}

\item{n_dims}{Number of dimensions to use for clustering and UMAP.}

\item{reduction}{Dimensionality reduction method to use (default is "pca").}

\item{clus}{Logical, if TRUE, clusters will be found using \code{FindNeighbors} and \code{FindClusters}.}

\item{viz}{Logical, if TRUE, UMAP will be run and visualized.}

\item{align}{Logical, if TRUE, clusters will be aligned between two fields.}

\item{field1}{First metadata field for alignment (default is "labels").}

\item{field2}{Second metadata field for alignment (default is "seurat_clusters").}

\item{field_aligned}{Field where aligned clusters are stored (default is "predicted_clusters").}

\item{params}{List of additional parameters, such as resolution for clustering and minimum distance for UMAP.}
}
\value{
The Seurat object with clustering and alignment information added.
}
\description{
This function clusters cells using PCA or other dimensional reduction methods and visualizes the results using UMAP.
Optionally aligns clusters between fields.
}
