% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{drgem_phaseI}
\alias{drgem_phaseI}
\title{Distributionally robust initialization}
\usage{
drgem_phaseI(
  X,
  meta,
  preprocess = T,
  scale.data = NULL,
  n_dims = 20,
  beta = 0.3,
  viz = T,
  var_features = NULL,
  align = T,
  field1 = "labels",
  field2 = "seurat_clusters",
  field_aligned = "weak_clusters",
  params1 = list(nfeatures = 200),
  params2 = list(res = 0.12, min.dist = 0.1, k.param = 20, return.model = F)
)
}
\arguments{
\item{X}{Seurat object to be processed.}

\item{meta}{Vector of class labels to subsample.}

\item{preprocess}{Logical, whether to reprocess the Seurat object during each iteration.}

\item{scale.data}{Number of iterations for subsampling.}

\item{n_dims}{Number of subsamples per class.}

\item{beta}{Number of principal components to use for PCA.}

\item{viz}{Logical, whether to visualize data (default = T)}

\item{var_features}{Number of features to use for variable feature selection.}

\item{align}{Logical, whether to align clusters during each iteration (default = T)}

\item{params}{List of additional parameters for clustering and UMAP.}
}
\value{
seurat object with weak labels
}
\description{
Use PCA as diagnostic tool for a distributionally robust initialization
}
