% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{drgem_phaseIII}
\alias{drgem_phaseIII}
\title{Perform Final Embedding}
\usage{
drgem_phaseIII(
  iobj,
  assignments,
  thres = 0.95,
  thres_by_class = NULL,
  n_dims = 20,
  n_features = 200,
  subsample_density = c(`1` = 800, `2` = 500, `3` = 450, `4` = 400, `5` = 50),
  params = list(res = 0.18, min.dist = 0.1, k.param = 10),
  return_ref = F
)
}
\arguments{
\item{iobj}{Seurat object to be processed.}

\item{assignments}{Dataframe with assignment labels and confidence scores.}

\item{thres}{Confidence threshold for selecting samples.}

\item{thres_by_class}{Class-specific confidence thresholds.}

\item{n_dims}{Number of principal components to use for the final embedding.}

\item{subsample_density}{List indicating the number of samples to subsample per class.}
}
\value{
The Seurat object with the final embedding and cluster visualization.
}
\description{
Performs final embedding on high-confidence subsampled cells.
}
