% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.OE}
\alias{get.OE}
\title{Overall Expression Wrapper}
\usage{
get.OE(r, sig)
}
\arguments{
\item{r}{data list}

\item{sig}{signature list}
}
\value{
overall expression matrix
}
\description{
wrapper around code that computes overall expression
from the given gene signature
}
