% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.OE1}
\alias{get.OE1}
\title{Overall Expression Subroutine}
\usage{
get.OE1(r, sig)
}
\arguments{
\item{r}{data list}

\item{sig}{signature list}
}
\value{
overall expression scores
}
\description{
code that computes overall expression
signature for each cell-type
}
