% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.tpm}
\alias{get.tpm}
\title{Function for computing "transcripts per million"}
\usage{
get.tpm(X, comp.reads)
}
\arguments{
\item{r}{data list}
}
\value{
r, your data list, with an r$tpm field added.
}
\description{
Further normalizes and transforms data to
common expression data output: "transcripts per million"
}
