% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%!in\%}
\alias{\%!in\%}
\title{Negated version of \code{\%in\%}}
\usage{
... \%!in\% NA
}
\arguments{
\item{x}{vector or NULL: the values to be matched.}

\item{table}{vector or NULL: the values to be matched against.}
}
\value{
A logical vector indicating if a match was found (negated).
}
\description{
This function returns a logical vector indicating if there are no matches or not.
}
