% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{initialize_pca}
\alias{initialize_pca}
\title{Initialize PCA and Seurat Object}
\usage{
initialize_pca(
  X,
  meta,
  preprocess = T,
  scale.data = NULL,
  n_dims = 30,
  params = list(nfeatures = 200),
  var_features = NULL
)
}
\arguments{
\item{X}{Matrix of gene expression counts.}

\item{meta}{Metadata for the cells. If NULL, only expression data will be used.}

\item{preprocess}{Logical, if TRUE, data will be normalized, variable features selected, and scaled.}

\item{scale.data}{Scaled data to be used directly if preprocessing is FALSE.}

\item{n_dims}{Number of principal components to compute.}

\item{params}{List of additional parameters, such as the number of features for variable feature selection (\code{nfeatures}).}

\item{var_features}{List of variable features, if predetermined}
}
\value{
A Seurat object with PCA results and reconstruction loss added to the metadata.
}
\description{
This function initializes a Seurat object from input data, optionally preprocesses it,
performs PCA, and calculates the reconstruction loss.
}
