% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{label_consensus}
\alias{label_consensus}
\title{Compute Consensus Label and Confidence}
\usage{
label_consensus(x, query)
}
\arguments{
\item{x}{A vector of cluster assignments, with possible values from the \code{query} or "UD" for unassigned.}
}
\value{
A named vector containing the consensus assignment, confidence score, and the number of votes.
}
\description{
This function computes the consensus label and confidence score for a set of cluster assignments.
}
