% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{map_clusters}
\alias{map_clusters}
\title{Map Clusters}
\usage{
map_clusters(a, b)
}
\arguments{
\item{a}{vector of cluster assignments from the first method}

\item{b}{vector of cluster assignments from the second method}
}
\value{
named vector indicating the mapped cluster assignments
}
\description{
Custom function that maps clusters between two cluster assignments based on
a contingency table; includes some edge cases.
}
