% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{map_clusters1}
\alias{map_clusters1}
\title{Map Clusters (Kuhn-Munkres)}
\usage{
map_clusters1(a, b)
}
\arguments{
\item{a}{vector of cluster assignments from the first method}

\item{b}{vector of cluster assignments from the second method}
}
\value{
named vector indicating the mapped cluster assignments
}
\description{
Function that maps clusters between two cluster assignments using Kuhn-Munkres
}
