% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot_cm}
\alias{plot_cm}
\title{Visualize Simple Plot}
\usage{
plot_cm(iobj, assignments, field1 = "labels", field2 = "assignment")
}
\arguments{
\item{iobj}{Seurat object}

\item{assignments}{data frame of consensus matrix from phase II!}

\item{field1}{string indicating the column name for the true labels}

\item{cluster_field}{string indicating the column name final predicted label}
}
\value{
None, saves the multiplot visualization to a file
}
\description{
Function that visualizes clustering results using multiple plots including PCA, UMAP, and reconstruction loss heatmap.
}
