% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot_sigs}
\alias{plot_sigs}
\title{Visualize Simple Plot}
\usage{
plot_sigs(fobj, features, ncol = 3)
}
\arguments{
\item{fobj}{Seurat object}

\item{features}{features to plot}

\item{ncol}{No. of columns of plots (default = 3)}
}
\value{
None, prints to console
}
\description{
Function that visualizes clustering results using multiple plots including PCA, UMAP, and reconstruction loss heatmap.
}
