% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prep4OE}
\alias{prep4OE}
\title{Prepocessing Function for Overall Expression}
\usage{
prep4OE(r, n.cat = 50)
}
\arguments{
\item{r}{data list}

\item{n.cat}{(default = 50)}
}
\value{
r, your data list
}
\description{
prepares data to compute overall expression
signature
}
