% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{subsample_by_class}
\alias{subsample_by_class}
\title{Sub-sample Cells by Class Label}
\usage{
subsample_by_class(query, obj, n_sub, field, seed = 888)
}
\arguments{
\item{query}{A vector of class labels to be sampled.}

\item{obj}{A Seurat object from which to sub-sample cells.}

\item{n_sub}{The number of cells to sub-sample per class.}

\item{seed}{A random seed to ensure reproducibility (default is 888).}
}
\value{
A vector of cell names that have been sub-sampled from the Seurat object.
}
\description{
This function sub-samples cells from a Seurat object by a specified class label, ensuring a consistent number of samples per class.
}
