% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{subsample_per_class}
\alias{subsample_per_class}
\title{Sub-sample Cells per Class Based on Density}
\usage{
subsample_per_class(assignments, cells, subsample_density, seed = 888)
}
\arguments{
\item{assignments}{A dataframe containing cell assignment labels.}

\item{cells}{A vector of cell names to be considered for sub-sampling.}

\item{subsample_density}{A named list indicating the number of cells to sub-sample for each class.}

\item{seed}{A random seed to ensure reproducibility (default is 888).}
}
\value{
A vector of sub-sampled cell names.
}
\description{
This function sub-samples cells based on a specified density for each class, considering assignments and cell counts.
}
