% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{t.test.mat}
\alias{t.test.mat}
\title{t-test for a matrix}
\usage{
\method{t}{test.mat}(m, b, two.sided = F, rankf = F, fold.changeF = T)
}
\arguments{
\item{m}{matrix}

\item{b}{boolean}

\item{two.sided}{whether to perform two-sided or not}

\item{rankf}{??}

\item{fold.changeF}{??}
}
\value{
pvalues in a table
}
\description{
performing t-tests on a per-row basis
}
