% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{transfer_embedding}
\alias{transfer_embedding}
\title{Transfer Embeddings Between Seurat Objects}
\usage{
transfer_embedding(
  sobj,
  tobj,
  n_dims = 10,
  semb = "pca",
  temb = "hrlpca",
  tkey = "hRLPC_",
  trecon = "recon_loss_hrlpca"
)
}
\arguments{
\item{sobj}{Source Seurat object from which to transfer PCA embeddings.}

\item{tobj}{Target Seurat object to which the embeddings are transferred.}

\item{n_dims}{Number of dimensions to use in the reconstruction.}

\item{semb}{Embedding reduction in the source object (default is "pca").}

\item{temb}{Embedding reduction in the target object (default is "hrlpca").}

\item{tkey}{Key for storing transferred embeddings in the target object.}

\item{trecon}{Field name for storing reconstruction loss in the target object.}
}
\value{
The updated target Seurat object with transferred embeddings and reconstruction loss.
}
\description{
Transfers embeddings from a source Seurat object to a target Seurat object.
}
