% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{visualize_multiplot}
\alias{visualize_multiplot}
\title{Visualize Multiplot}
\usage{
visualize_multiplot(
  obj,
  name = "",
  recon_field = "recon_loss",
  cluster_field = "seurat_clusters",
  pub = T,
  custom.colors = NULL,
  height = 8,
  width = 12
)
}
\arguments{
\item{obj}{Seurat object}

\item{name}{string for naming the output file}

\item{recon_field}{string indicating the column name for reconstruction loss (default = "recon_loss")}

\item{cluster_field}{string indicating the column name for cluster assignments (default = "seurat_clusters")}
}
\value{
None, saves the multiplot visualization to a file
}
\description{
Function that visualizes clustering results using multiple plots including PCA, UMAP, and reconstruction loss heatmap.
}
