% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{visualize_simpleplot}
\alias{visualize_simpleplot}
\title{Visualize Simple Plot}
\usage{
visualize_simpleplot(
  obj,
  true_labels = "labels",
  weak_labels = "seurat_clusters",
  custom.colors = NULL,
  type1 = "discrete",
  type2 = "discrete",
  reduction = "umap",
  midpoint = NULL
)
}
\arguments{
\item{obj}{Seurat object}

\item{true_labels}{string indicating the column name for reconstruction loss (default = "recon_loss")}

\item{custom.colors}{list of custom color mappings (default = NULL)}

\item{predicted_labels}{string indicating the column name for cluster assignments (default = "seurat_clusters")}
}
\value{
None, prints plots to console
}
\description{
Function that visualizes clustering results using multiple plots including PCA, UMAP, and reconstruction loss heatmap.
}
